src/requirements.txt
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
colorama==0.4.4
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.6
django-allauth==0.45.0
django-autoslug==1.9.8
django-debug-toolbar==3.2.1
django-environ==0.4.5
django-extensions==3.1.3
django-filter==2.4.0
django-jazzmin==2.4.8
django-redis-cache==2.1.0
django-rq==2.1.0
django-tinymce==3.3.0
django-translation-manager==1.1.1
djangorestframework==3.12.4
environ==1.0
idna==3.2
ipython==7.26.0
ipython-genutils==0.2.0
jedi==0.18.0
matplotlib-inline==0.1.2
oauthlib==3.1.1
parso==0.8.2
pickleshare==0.7.5
Pillow==8.3.1
polib==1.1.1
prompt-toolkit==3.0.19
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
rq==1.9.0
six==1.16.0
sqlparse==0.4.1
traitlets==5.0.5
urllib3==1.26.6
wcwidth==0.2.5
